### 2020-08-26 #2600 @whedon
[PRE REVIEW]: RCaller: A Java package for interfacing R
**Submitting author:** @jbytecode (<a href="http://orcid.org/0000-0002-9402-1982">Mehmet Hakan Satman</a>)
**Repository:** <a href="https://github.com/jbytecode/rcaller/" target ="_blank">https://github.com/jbytecode/rcaller/</a>
**Version:** v3.0
