+++ paper/paper.bib
@@ -3,6 +3,7 @@
   author={Satman, M Hakan},
   journal={Journal of Advances in Mathematics and Computer Science},
   pages={2188--2196},
+  doi={10.9734/bjmcs/2014/10902},
   year={2014}
 }
 
@@ -31,6 +32,7 @@
   number={2},
   pages={303--311},
   year={2009},
+  doi={10.1007/s00180-008-0132-x},
   publisher={Springer}
 }
 
@@ -46,6 +48,7 @@
   author={Niephaus, Fabio and Felgentreff, Tim and Hirschfeld, Robert},
   booktitle={Proceedings of the Conference Companion of the 3rd International Conference on Art, Science, and Engineering of Programming},
   pages={1--3},
+  doi={10.1145/3328433.3328458},
   year={2019}
 }
 
