### 2020-08-28 e514781
some documentation on RCaller.java
