+++ tasks.md
@@ -1,20 +1,0 @@
-####### changelog for version 3.0 #######
-
-- [X] javax.script interface for RCaller
-- [X] *R* command options
-- [X] Core *R* makexml() enhancements (removing some loop, adding apply family functions)
-- [X] Improved *R* and *Rscript* detection facilities
-- [X] Fixed stop operation for RStreamHandler
-- [X] Some UTF-8 support
-- [X] XML special characters such as less-than, greater-than, ampersand, and apostrophe in variable names are handled
-- [X] static RCaller.create() factory methods are implemented for handling more options
-- [X] Get all issues solved and closed in github
-- [ ] Add new examples for new features
-- [X] Add ~steel~ tests for a more reliable software library
-- [X] Solve all of the issues, remove bugs due to problems, at least give some advice in stackoverflow
-- [X] Passing fields of Java Objects to R in a more clear way
-- [X] Invocable interface is implemented for R function call facility from Java in RCallerScriptEngine
-- [ ] Threading improvements in stream consumers
-- [X] Minimum factor data type support
-- [X] Minimum data.frame support
-
