+++ README.md
@@ -24,7 +24,7 @@ suitable for the projects which have many clients that request relatively single
 
 # Usage
 
-It is recommended to take a look at the existing [tests](https://github.com/jbytecode/rcaller/tree/master/RCaller/src/test/java/org/expr/rcaller)
+It is recommended to take a look at the existing [tests](https://github.com/jbytecode/rcaller/tree/master/RCaller/src/test/java/com/github/rcaller)
  and 
 [examples](https://github.com/jbytecode/rcaller/tree/master/RCaller/src/main/java/examples)
 
