### 2020-08-28 cf8e0ab
stringbuffer type changed to stringbuilder in private code of RCode
