+++ RCaller/src/main/java/com/github/rcaller/util/RCodeUtils.java
@@ -39,31 +39,31 @@ import java.util.logging.Logger;
 
 public class RCodeUtils {
 
-    public static void addIntArray(StringBuffer rCode, String name, int[] arr, boolean useEquals) {
+    public static void addIntArray(StringBuilder rCode, String name, int[] arr, boolean useEquals) {
         addArray(rCode, name, ArrayUtils.toObject(arr), useEquals, false);
     }
 
-    public static void addLongArray(StringBuffer rCode, String name, long[] arr, boolean useEquals) {
+    public static void addLongArray(StringBuilder rCode, String name, long[] arr, boolean useEquals) {
         addArray(rCode, name, ArrayUtils.toObject(arr), useEquals, false);
     }
 
-    public static void addFloatArray(StringBuffer rCode, String name, float[] arr, boolean useEquals) {
+    public static void addFloatArray(StringBuilder rCode, String name, float[] arr, boolean useEquals) {
         addArray(rCode, name, ArrayUtils.toObject(arr), useEquals, false);
     }
 
-    public static void addDoubleArray(StringBuffer rCode, String name, double[] arr, boolean useEquals) {
+    public static void addDoubleArray(StringBuilder rCode, String name, double[] arr, boolean useEquals) {
         addArray(rCode, name, ArrayUtils.toObject(arr), useEquals, false);
     }
 
-    public static void addStringArray(StringBuffer rCode, String name, String[] arr, boolean useEquals) {
+    public static void addStringArray(StringBuilder rCode, String name, String[] arr, boolean useEquals) {
         addArray(rCode, name, arr, useEquals, true);
     }
 
-    public static void addShortArray(StringBuffer rCode, String name, short[] arr, boolean useEquals) {
+    public static void addShortArray(StringBuilder rCode, String name, short[] arr, boolean useEquals) {
         addArray(rCode, name, ArrayUtils.toObject(arr), useEquals, false);
     }
 
-    public static void addLogicalArray(StringBuffer rCode, String name, boolean[] arr, boolean useEquals) {
+    public static void addLogicalArray(StringBuilder rCode, String name, boolean[] arr, boolean useEquals) {
         String[] stringArray = new String[arr.length];
         for (int i = 0; i < arr.length; i++) {
             stringArray[i] = String.valueOf(arr[i]).toUpperCase();
@@ -71,7 +71,7 @@ public class RCodeUtils {
         addArray(rCode, name, stringArray, useEquals, false);
     }
 
-    public static <T> void addArray(StringBuffer rCode, String name, T[] array, boolean useEquals, boolean isString) {
+    public static <T> void addArray(StringBuilder rCode, String name, T[] array, boolean useEquals, boolean isString) {
         if (!name.equals("")) {
             if (useEquals) {
                 rCode.append(name).append("=");
@@ -98,14 +98,14 @@ public class RCodeUtils {
         }
     }
 
-    public static void addJavaObject(StringBuffer rCode, Object o, boolean useEquals) throws IllegalAccessException {
+    public static void addJavaObject(StringBuilder rCode, Object o, boolean useEquals) throws IllegalAccessException {
         rCode.append(((JavaObject) o).produceRCode(useEquals));
         if (!useEquals) {
             rCode.append("\n");
         }
     }
 
-    public static void addDoubleMatrix(StringBuffer rCode, String name, double[][] matrix, boolean useEquals) {
+    public static void addDoubleMatrix(StringBuilder rCode, String name, double[][] matrix, boolean useEquals) {
         int dim2 = matrix[0].length;
         int counter = 0;
         if (!name.equals("")) {
@@ -131,31 +131,31 @@ public class RCodeUtils {
         }
     }
 
[…]
