### 2020-08-28 47b8dff
Create CONTRIBUTING.md
