+++ CONTRIBUTING.md
@@ -1,0 +1,31 @@
+# Welcome!
+
+Our contributers are mainly RCaller users in academia or software sector. They used to integrate RCaller in their projects and they contributed to the project 
+for supporting themself first and the whole users after. You can contribute by
+
+- sending pull requests (code is the king)
+- writing blogposts
+- making youtube videos
+- posting answers in stackoverflow.
+
+## Sending pull requests
+If you added some functionality, tests, or examples in the project, first create a fork in Github and make the contribution public in your repo. 
+Then send a pull request as decribed in [Creating pull requests](https://docs.github.com/en/github/collaborating-with-issues-and-pull-requests/creating-a-pull-request).
+
+## Writing blogposts
+Many users have blogposts about RCaller. If you have one, tell us to be visible in this repo. Here is some blogposts:
+- [Practical code solutions](http://stdioe.blogspot.com/search/label/rcaller).
+- [Scott Hoover] (https://scottdhoover.wordpress.com/2013/05/08/a-basic-rcaller-example/).
+- [Baeldung] (https://www.baeldung.com/java-r-integration).
+- [Renjin] (http://docs.renjin.org/en/latest/introduction.html).
+
+## Making youtube videos
+Videos are perfect tools for tutorials. Here is some examples:
+- [video by Bryan Downing - 1] (https://www.youtube.com/watch?v=hZDwuCV2fO8).
+- [video by Bryan Downing - 2] (https://www.youtube.com/watch?v=fHt0cq-6BvM).
+
+## Posting answers in stackoverflow
+Stackoverflow is the defacto forum for nearly all tools in area of software. RCaller questions can be found using the tag **rcaller**.
+- [Questions on stackoverflow] (https://stackoverflow.com/questions/tagged/rcaller).
+
+We are just here waiting to hear from you!
