### 2020-08-28 dda6fcd
Create CODE_OF_CONDUCT.md
