### 2020-08-28 #2600 @jbytecode
since this repo is moved from old [code.google.com](https://code.google.com/archive/p/rcaller/source/default/commits), I am represented in above table by more than one nicknames, so 'hako', 'jbytecode','Mehmet Hakan Satman', and 'mhsatman' are both I am. fyi.
