### 2020-09-02 #2600 @Kevin-Mattheus-Moerman
@jbytecode

Some notes on your paper:
- [ ] Typo: `persfective` -> `perspective`

- [ ] We do not accept previously published scholarly works in JOSS. If major changes have been made, or if the focus on the paper is significantly different, exceptions can be made. Therefore, can you please clarify what the difference in focus/content/intent is for the current submission with respect to this paper: `Satman, M. H. (2014). RCaller: A software library for calling r from java.Journal of Advances in Mathematics and Computer Science, 2188–2196. doi:10.9734/bjmcs/2014/10902`. Likely you are looking to submit for the new developments from the past 6 years but please formally clarify the difference and list some of these developments please.
