### 2020-09-02 #2600 @jbytecode
@Kevin-Mattheus-Moerman Dear editor, new features such as support for DataFrames, R Start-up options, Java Scripting API (JSR 223), performance improvements, and minor API changes are not included in the paper you mentioned above. I will highlight the differences. Thanks.
