### 2020-09-02 d1faef5
typo correction (persfective -> perspective)
