+++ paper/paper.md
@@ -17,7 +17,7 @@ bibliography: paper.bib
 ---
 
 # Summary
-Nowadays, the interest on statistical computation and data analysis increased the popularity of some programming languages such as R, Python, and Julia. As a result of popularity, hundreds of packages are developed for these targeted languages. On the other side, many mainframe languages are adopted for more specific tasks in server-side or the general-purpose use in several platforms. Interaction between those languages became a major requirement due to the lack of modern statistical tools written in the general-purpose languages such as Java. Renjin [@renjin] is a great effort for interfacing R from within Java using a different persfective: Writing R from scratch in Java! GraalVM [@graalvm] is an other admirable attempt for interfacing Python and R in Java with its polyglot design. Many tools and software packages are developed for interfacing R and Java. Each one stands out with its different features including speed, installation procedures, learning speed, and ease of use. 
+Nowadays, the interest on statistical computation and data analysis increased the popularity of some programming languages such as R, Python, and Julia. As a result of popularity, hundreds of packages are developed for these targeted languages. On the other side, many mainframe languages are adopted for more specific tasks in server-side or the general-purpose use in several platforms. Interaction between those languages became a major requirement due to the lack of modern statistical tools written in the general-purpose languages such as Java. Renjin [@renjin] is a great effort for interfacing R from within Java using a different perspective: Writing R from scratch in Java! GraalVM [@graalvm] is an other admirable attempt for interfacing Python and R in Java with its polyglot design. Many tools and software packages are developed for interfacing R and Java. Each one stands out with its different features including speed, installation procedures, learning speed, and ease of use. 
 
 
 # Statement of need 
