### 2020-09-02 ed2b2af
Update paper.md
