### 2020-09-10 #2600 @mikldk
@jbytecode Can you please explain/clarify why the other (major) committers are not co-authors (maybe they declined?)?
