### 2020-09-10 jbytecode/rcaller#42 @jbytecode
Manuscript submission
Hi friends,

I have prepared a manuscript and submitted to Journal of Open Source Software that presents RCaller. Since you all have valuable contributions, the dear editor @mikldk kindly asked me to ask if you would like to participate in this article. Here is the pre-review link: [https://github.com/openjournals/joss-reviews/issues/2600](https://github.com/openjournals/joss-reviews/issues/2600)

If your answer is yes, please add your contributions and your name to the paper hosted in our RCaller repository ([the paper directory](https://github.com/jbytecode/rcaller/tree/master/paper)). If your answer is no, please make an explanation.

Best wishes.

@paulcurcean
@mbatchkarov ✓
@Kopilov ✓
@dieresys ✓
@wongjoel ✓
@ksfx ✓
@waffle-iron
@stevensotelo
@gitter-badger
@edineipiovesan
@sjcetx ✓
