### 2020-09-11 jbytecode/rcaller#42 @wongjoel
Well, my contributions were to help out with making the R path on Windows more flexible. I don't think that makes me a major committer, so I'm happy with just being in the Acknowledgement section.
