### 2020-09-11 jbytecode/rcaller#42 @ksfx
I only made tiny contributions. I'm very happy with the Acknowledgement section
