### 2020-09-11 jbytecode/rcaller#42 @Kopilov
I have added StopRCallerAsync method that allows to kill long or freezed calculation from separate thread (actually, for timeout feature asked by customer in my my project https://github.com/Kopilov/RCallerService)
In the process "TODO checking file length is wrong" had been fixed.

@jbytecode, would you like me to commit something to https://github.com/jbytecode/rcaller/tree/master/paper directory?
I do not have any experience with JOSS publishing, sorry
