### 2020-09-11 jbytecode/rcaller#42 @jbytecode
@Kopilov Since you have contributions to RCaller, you have rights to be a co-author in the manuscript. If you want to be a part of this manuscript, you can make contributions to the md file in the repo and add your name. I can review your changes and try to bring it to the shape required by the journal. JOSS has a well-defined list of publishing rules, so you can carefully read it. The submitted paper is a good starting point.
