### 2020-09-11 jbytecode/rcaller#42 @jbytecode
@Kopilov as an opinion, maybe you can refer to RCallerService, so we can show up that the RCaller is not just a stand-alone library, it is used in another libraries as a base.
