### 2020-09-11 jbytecode/rcaller#43 @jbytecode
Contributions to manuscript prepared for JOSS
We can discuss possible contributions to the manuscript here. @Kopilov will you participate? I can help to you if you have any confusions.
