### 2020-09-11 jbytecode/rcaller#43 @Kopilov
@jbytecode I am ready, thanks!

At first, is orcid necessary to add me to the authors list? If yes, how can I get it?

My single scientific work is xAPI translation to russian (started as https://github.com/Kopilov/xAPI-Spec/blob/translate_ru/xAPI_ru.md, published as http://books.ifmo.ru/file/pdf/1772.pdf)
