### 2020-09-11 jbytecode/rcaller#43 @jbytecode
Yes you can take a new researcher id. Please, open the page https://orcid.org/ and register.

No previous scientific work is required, you have contributions in RCaller, that is enough.
