### 2020-09-11 jbytecode/rcaller#42 @jbytecode
We have a new issue here for possible questions and discussions for those who accept to participate.

[https://github.com/jbytecode/rcaller/issues/43](https://github.com/jbytecode/rcaller/issues/43)
