### 2020-09-11 jbytecode/rcaller#43 @Kopilov
Commited to https://github.com/Kopilov/rcaller/blob/paper/paper/paper.md
The table is not displayed as a table anymore :(
My patch is https://github.com/Kopilov/rcaller/commit/a8a2cbbc4f03d6dfe3750577e7007377a524ab5b
