### 2020-09-11 jbytecode/rcaller#43 @Kopilov
Nevertheless, what should be put in "affiliation" field?
