### 2020-09-11 jbytecode/rcaller#43 @Kopilov
@jbytecode yes, thanks!
I have fixed the last paragraph, now will try to make correct reference by the bibtex file
