### 2020-09-11 jbytecode/rcaller#43 @jbytecode
as a note for all contributers:

when a reference is used like [@SomeCitationsHereAsAKeyword] in md file, an entry should be placed like

@misc{SomeCitationsHereAsAKeyword,
      title        = "Pluto: The 'Other' Red Planet",
      author       = "{NASA}",
      howpublished = "\url{https://www.nasa.gov/nh/pluto-the-other-red-planet}",
      year         = 2015,
      note         = "Accessed: 2018-12-06"
    }

in bibtex file. If the entry type is a paper, a doi number is required.
