+++ paper/paper.bib
@@ -62,3 +62,8 @@
   year={2013}
 }
 
+@manual{rcallerservice,
+  title={RCallerService},
+  author={Aleksandr Kopilov},
+  url={https://github.com/Kopilov/RCallerService}
+}
