+++ paper/paper.bib
@@ -65,5 +65,6 @@
 @manual{rcallerservice,
   title={RCallerService},
   author={Aleksandr Kopilov},
+  year={2020},
   url={https://github.com/Kopilov/RCallerService}
 }
