+++ paper/paper.bib
@@ -63,7 +63,7 @@
 }
 
 @manual{rcallerservice,
-  title={RCallerService},
+  title={RCallerService — microservice for running R-scripts by HTTP},
   author={Aleksandr Kopilov},
   year={2020},
   url={https://github.com/Kopilov/RCallerService}
