### 2020-09-13 jbytecode/rcaller#42 @dieresys
I only made tiny contributions. I'm very happy with the Acknowledgement section.
Thanks!
