### 2020-09-15 jbytecode/rcaller#42 @jbytecode
@ sjcetx welcome to RCaller. Since you are a new contributor, please read the entire issue and give a response. Thank you in advance. If your answer is no, please write your name so we can put it in the Acknowledge section.
