### 2020-09-15 jbytecode/rcaller#42 @sjcetx
As my contributions were minimal being in the acknowledgements is more than enough.
Thanks,
Simon Carter
