### 2020-09-15 jbytecode/rcaller#42 @jbytecode
> As my contributions were minimal being in the acknowledgements is more than enough.
> Thanks,
> Simon Carter

okay, I added your name in the  acknowledgements part of manuscript @sjcetx . Keep with us.
