+++ .github/workflows/maven.yml
@@ -1,24 +1,0 @@
-# This workflow will build a Java project with Maven
-# For more information see: https://help.github.com/actions/language-and-framework-guides/building-and-testing-java-with-maven
-
-name: Java CI with Maven
-
-on:
-  push:
-    branches: [ master ]
-  pull_request:
-    branches: [ master ]
-
-jobs:
-  build:
-
-    runs-on: ubuntu-latest
-
-    steps:
-    - uses: actions/checkout@v2
-    - name: Set up JDK 13
-      uses: actions/setup-java@v1
-      with:
-        java-version: 13
-    - name: Build with Maven
-      run: cd RCaller; mvn -B package --file pom.xml
