+++ .github/workflows/maven_test.yml
@@ -1,0 +1,29 @@
+# This workflow will build a Java project with Maven
+# For more information see: https://help.github.com/actions/language-and-framework-guides/building-and-testing-java-with-maven
+
+name: test
+
+on:
+  push:
+    branches: [ master, citest ]
+  pull_request:
+    branches: [ master ]
+
+jobs:
+
+  test:
+
+    runs-on: ubuntu-18.04
+
+    steps:
+    - uses: actions/checkout@v2
+    - name: Set up JDK 1.8
+      uses: actions/setup-java@v1
+      with:
+        java-version: 1.8
+    - name: Set up R
+      run: |
+        sudo apt-get update
+        sudo apt-get -y install r-base
+    - name: Build with Maven
+      run: cd RCaller && mvn test
