### 2020-09-21 #2600 @jbytecode
Dear editor @mikldk , after asking all contributors to join manuscript 11 days ago, one of the major contributor @Kopilov joined, other major contributor is no longer interested in the project and has no respond. The third major one and some of other minor contributors are happy being in the Acknowledgement section, some of them also have no response. Any actions needed?
