### 2020-09-23 jbytecode/rcaller#42 @edineipiovesan
I fixed some charset issues and added default path for R in macOS. Pretty tiny changes, so I'll be happy with the Acknowledgement section.
Thanks,
Edinei Piovesan
