### 2020-09-26 07e5c92
Explicitly set locale en_US.UTF-8 for tests
