+++ .github/workflows/maven_test.yml
@@ -25,5 +25,8 @@ jobs:
       run: |
         sudo apt-get update
         sudo apt-get -y install r-base
+        sudo locale-gen "en_US.UTF-8"
+        sudo bash -c 'echo "LANG=en_US.UTF-8" > /etc/default/locale'
+        echo '::set-env name=LANG::en_US.UTF-8'
     - name: Build with Maven
       run: cd RCaller && mvn test
