### 2020-09-26 7397a90
Remove citest branch from yaml, revert jbytecode in badges
