+++ .github/workflows/maven_test.yml
@@ -5,7 +5,7 @@ name: test
 
 on:
   push:
-    branches: [ master, citest ]
+    branches: [ master ]
   pull_request:
     branches: [ master ]
 
