### 2020-10-17 jbytecode/rcaller#43 @nthomasCUBE
Hi,

I quess in the example from your manuscript, you have defined "en" but then use "engine" (e.g. engine.put("a",a)
like here:

		ScriptEngineManager manager=new ScriptEngineManager();
		ScriptEngine **en**=manager.getEngineByName("RCaller");

		System.out.println("JA");

		try {
			double[] a=new double[] {19.0, 17.0, 23.0};
			**en**.put("a",a);
			**en**.eval("sortedA <- sort(a)");
			double[] result=(double[])en.get("sortedA");
			System.out.println(result[0]);
		} catch (ScriptException e) {
			// TODO Auto-generated catch block
			e.printStackTrace();
		}

br
