### 2020-10-17 #2722 @nthomasCUBE
added new issue for updating example in the manuscript
https://github.com/jbytecode/rcaller/issues/43#issuecomment-710939070
