### 2020-10-18 febe465
Revert "problematic test UTF8 commented for MacOs builds"
This reverts commit 79221e61cc9d78f4d4faad9214f060f01e3f9432.
