### 2020-10-18 30e10e8
Add no-lang standard locales support
