### 2020-10-31 jbytecode/rcaller#54 @xin-huang
[JOSS REVIEW] Suggestion for Software paper
This is a part of the JOSS review (openjournals/joss-reviews#2722)

1. For the title, I suggest the authors add the version into the title as [their manual](https://github.com/jbytecode/rcaller/blob/master/doc/rcaller3/rcaller3.pdf), i.e. RCaller 3.0: A Java package for interfacing R
2. I feel the Summary section is more like a state of the field or introduction. I suggest the authors change this section to a state of the field, as the review checklist suggests.
3. I suggest the authors write a new summary, for example, the abstract in [the manual](https://github.com/jbytecode/rcaller/blob/master/doc/rcaller3/rcaller3.pdf) may be a good summary for the paper.
4. The target audience is not clear in the statement of need, I suggest the author make it clear as the review checklist suggests.
