### 2020-10-31 jbytecode/rcaller#55 @xin-huang
 [JOSS REVIEW] Suggestion for Documentation
This is a part of the JOSS review (openjournals/joss-reviews#2722)

1. It is not clear that who is the target audience, I suggest the authors make it clear as the review checklist suggests.
2. For the installation, the version of RCaller in the Maven dependency code in [the manual](https://raw.githubusercontent.com/jbytecode/rcaller/master/doc/rcaller3/rcaller3.pdf) is 3.0, whereas it is 2.8 in the README.md. This makes me confused which version I should use. I also feel that the instruction for installation in the manual is better than that in the README.md. It tells readers more clearly that there are two approaches to install RCaller, one is through Maven, the other is by downloading the jar file and add it into the classpath. I suggest the authors also replace the instruction for installation in the README.md by the instruction in the manual.
3. It is not clear that what are the dependencies for RCaller, at least, the authors should state which version of Java is required for RCaller. Or does RCaller support all the versions of Java? Based on my simple test, it looks like RCaller does not support Java 1.5. Besides, RCaller is an interface between R and Java, therefore, another dependency should be R at least. Does RCaller support all the versions of R? In Page 3 of the manual, the second paragraph of the Setup and Installation section, it says that all the dependencies including the libraries that RCaller imports are downloaded. What are the dependencies that RCaller imports and downloads? I suggest the authors make the versions of Java and R as well as the dependencies clear in the documentation. Also, I suggest the authors list the dependencies in the README.md for the RCaller repo, so that users do not need to search for them in the long manual.
4. The authors provide a lot of [examples](https://github.com/jbytecode/rcaller/tree/master/RCaller/src/main/java/examples), but it is not clear that what the purpose for each example is. For example, the authors only named the examples such as Example2.java, Example3.java and so on. It is difficult for users to figure out what the example is used for by themselves. I suggest the authors rename the example files and add annotation into the examples for users to have a better understanding. Also, I suggest the authors list their examples in the README.md, so that users can quickly find out the examples they need instead of searching examples one by one.
The authors also provided several code examples in [the manual](https://raw.githubusercontent.com/jbytecode/rcaller/master/doc/rcaller3/rcaller3.pdf), e.g., Listing 3. However, after copying these codes and pasting them into the IDE, these codes became weird and I have to reformat them. Therefore, I suggest the authors use more simple style, e.g. remove the line number and use plain text, for their code examples in the manual, so that users can copy and test the codes more easily.
5. API method documents are missing. The authors should add them into the repo as the review checklist suggests.
6. The community guidelines are missing. The authors should add them into the README.md as the review checklist suggests.
