### 2020-10-31 43f88de
refactor examples
