+++ RCaller/src/main/java/examples/DescriptiveStatistics.java
@@ -1,27 +1,3 @@
-/*
- *
-RCaller, A solution for calling R from Java
-Copyright (C) 2010,2011  Mehmet Hakan Satman
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU Lesser General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU Lesser General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- *
- * Mehmet Hakan Satman - mhsatman@yahoo.com
- * http://www.mhsatman.com
- * Google code projec: https://github.com/jbytecode/rcaller
- *
- */
 package examples;
 
 import com.github.rcaller.rstuff.RCaller;
@@ -37,7 +13,7 @@ import java.util.logging.Logger;
  * @since 2.0
  * @version 2.0
  */
-public class Example3 {
+public class DescriptiveStatistics {
 
   /**
    * Standalone test application.
@@ -45,7 +21,7 @@ public class Example3 {
    * from Java.
    *
    */
-  public Example3() {
+  public DescriptiveStatistics() {
     try {
 
       /**
@@ -142,12 +118,12 @@ public class Example3 {
         System.out.print(result + ", ");
       }
     } catch (Exception e) {
-      Logger.getLogger(Example3.class.getName()).log(Level.SEVERE, e.getMessage());
+      Logger.getLogger(DescriptiveStatistics.class.getName()).log(Level.SEVERE, e.getMessage());
     }
   }
   
   public static void main(String[] args) {
-    new Example3();
+    new DescriptiveStatistics();
   }
 }
 
