### 2020-10-31 ad8245c
refactor forecasting example
