### 2020-10-31 332af1a
add examples and dependencies in README.md
