### 2020-10-31 #2722 @xin-huang
The authors implemented an interesting tool for Java programmers to use R codes inside Java, which may be helpful for Java programmers to use tremendous statistical tools in R. However, I feel the authors could improve their documentation and software paper and make them more user-friendly. I have open two issues (https://github.com/jbytecode/rcaller/issues/54 and https://github.com/jbytecode/rcaller/issues/55) in the repo. I hope the authors could address my concerns.
Besides, although the authors did not mention performance in the software paper, they discussed this issue in [the manual](https://raw.githubusercontent.com/jbytecode/rcaller/master/doc/rcaller3/rcaller3.pdf). Therefore, I suggest the authors release their codes for the benchmark, so that users can verify the performance issue more easily.
