+++ doc/rcaller3/rcaller3.tex
@@ -18,6 +18,7 @@
 \usepackage{url}
 \usepackage{float}
 \usepackage{menukeys}
+\usepackage[nottoc,notlot,notlof]{tocbibind}
 
 
 
@@ -86,10 +87,6 @@ In this paper, we introduce the version $3.0$ of the software library \texttt{RC
   basicstyle=\ttfamily,
   frame=single,
   morekeywords={version,dependency,groupId,artifactId,classifier,c}
-  numbers=left,                    
-  numbersep=5pt,                   
-  numberstyle=\tiny\color{color_black},
-  numberfirstline=true,
   backgroundcolor=\color{color_code_back_gray}
 }
 \begin{lstlisting}[caption=Downloading source code, label=code:git]
@@ -106,10 +103,6 @@ The most straightforward way of including \texttt{RCaller} in a project as a dep
   basicstyle=\ttfamily,
   frame=single,
   morekeywords={version,dependency,groupId,artifactId,classifier,c}
-  numbers=left,                    
-  numbersep=5pt,                   
-  numberstyle=\tiny\color{color_black},
-  numberfirstline=true,
   backgroundcolor=\color{color_code_back_gray}
 }
 \begin{lstlisting}[caption=Maven dependency code, label=code:maven_dependency]
@@ -140,12 +133,7 @@ The class \texttt{RCaller} includes methods for creating external processes and 
 \lstset{
   language=Java,
   showstringspaces=false,
-  numbers=left,
   frame=single,
-  numbers=left,                    
-  numbersep=5pt,                   
-  numberstyle=\tiny\color{color_black},
-  numberfirstline=true,
   backgroundcolor=\color{color_code_back_gray}
 }
 \begin{lstlisting}[caption=Simple call,label=code_simple_call]
@@ -185,17 +173,12 @@ In Listing \ref{code_simple_call}, the mean of an vector is calculated on \textt
 \noindent where the arguments for variable names in type of \texttt{String} and values in corresponding type. Listing \ref{code_java_pass_data} shows an example of passing different types of data to \texttt{R} and getting back the results in \texttt{Java}.
 
 \begin{minipage}{\linewidth}
-\lstset{
-  language=Java,
-  numbers=left,
-  frame=single,
-  numbers=left,                    
-  numbersep=5pt,                   
-  numberstyle=\tiny\color{color_black},
-  numberfirstline=true,
-  showstringspaces=false,
-  backgroundcolor=\color{color_code_back_gray} 
-}
+  \lstset{
+    language=Java,
+    showstringspaces=false,
+    frame=single,
+    backgroundcolor=\color{color_code_back_gray}
+  }
 \begin{lstlisting}[caption=Passing Data,label=code_java_pass_data]
 RCode code = RCode.create();
 RCaller caller = RCaller.create();
@@ -234,17 +217,12 @@ In addition to array-passing functions, the function \texttt{addDoubleMatrix} ca
 
 
 \begin{minipage}{\linewidth}
-\lstset{
-  language=Java,
-  numbers=left,
-  frame=single,
-  numbers=left,                    
-  numbersep=5pt,                   
-  numberstyle=\tiny\color{color_black},
-  numberfirstline=true,
-  showstringspaces=false,
-  backgroundcolor=\color{color_code_back_gray} 
-}
+  \lstset{
+    language=Java,
+    showstringspaces=false,
+    frame=single,
+    backgroundcolor=\color{color_code_back_gray}
+  }
 \begin{lstlisting}[caption=Transferring Matrices,label=code_java_svd]
 RCaller caller = RCaller.create();
 RCode code = RCode.create();
@@ -268,17 +246,12 @@ double[][] v = caller.getParser().getAsDoubleMatrix("v");
 
 
 \begin{minipage}{\linewidth}
-\lstset{
[…]
