+++ RCaller/target/apidocs/package-search-index.js
@@ -1,0 +1,1 @@
+packageSearchIndex = [{"l":"All Packages","u":"allpackages-index.html"},{"l":"com.github.rcaller"},{"l":"com.github.rcaller.datatypes"},{"l":"com.github.rcaller.exception"},{"l":"com.github.rcaller.graphics"},{"l":"com.github.rcaller.io"},{"l":"com.github.rcaller.rstuff"},{"l":"com.github.rcaller.scriptengine"},{"l":"com.github.rcaller.statistics"},{"l":"com.github.rcaller.util"},{"l":"examples"}];updateSearchResults();
