### 2020-10-31 jbytecode/rcaller#55 @jbytecode
Dear viewer @xin-huang

1. "Who uses RCaller?" section added in [README.md](https://github.com/jbytecode/rcaller#who-uses-rcaller)

2. Instructions for installing and building is added in [README.md](https://github.com/jbytecode/rcaller#building-and-installing-rcaller)

3. Dependencies section added in [README.md](https://github.com/jbytecode/rcaller#dependencies). RCaller compiles with JDK1.8 but I don't remember what was the oldest revision of R was tested. There is a version history of R [here](https://github.com/r-hub/rversions) and I think v2.0.0 or higher is enough. (Today v4.x.x is current).
However, we don't use any version specific functionality of R. This is explained in the section.

4. The manual LaTeX file is updated, a new pdf is compiled, line numbers removed for easy copy-paste stuff. pdf is [here](https://github.com/jbytecode/rcaller/blob/master/doc/rcaller3/rcaller3.pdf).
Examples are refactored by means of file names, descriptions and comments. Some examples are suitable to ship in a single file so the number of examples is changed. A selected subset of examples are presented in [README.md](https://github.com/jbytecode/rcaller#usage)

5. JavaDocs are compiled and hosted [here](https://github.com/jbytecode/rcaller/tree/master/RCaller/target/apidocs).

6. We have [a page](https://github.com/jbytecode/rcaller/blob/master/CONTRIBUTING.md) for newcomers. It is added to [README.md](https://github.com/jbytecode/rcaller/blob/master/CONTRIBUTING.md).

I corrected & added & removed all the stuff you mentioned above. Please specify if there is anything I forgot or misunderstood.

Best regards.
