### 2020-10-31 jbytecode/rcaller#55 @jbytecode
Dear reviewer @xin-huang ,

Based on your comment

> Besides, although the authors did not mention performance in the software paper, they discussed this issue in the manual. Therefore, I suggest the authors release their codes for the benchmark, so that users can verify the performance issue more easily.

in JOSS review page, I regret to state that the code we used in the benchmark is not available. If you find it really necessary to present time consumed by the API, I can replicate the simulations using the data generating process given in the manual paper. However this may have time.

Regards.
