+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'RCaller: A Java package for interfacing R'
+title: 'RCaller 3.0: A Java package for interfacing R'
 tags:
   - R
   - Java
@@ -22,11 +22,14 @@ bibliography: paper.bib
 ---
 
 # Summary
+In this paper, version 3.0 of the open source library RCaller is introduced. RCaller is a software library which simplifies performing data analysis and statistical calculations in Java using R. The details are hidden from the user including transferring data between platforms, function calls, and retrieving results. Addition to the previous revisions, RCaller 3.0 implements the scripting API of Java in which the R function calls and data transfers are performed in a standard way as in the way of calling other scripting languages in Java. Besides implementation of new features, RCaller has many performance improvements in the new release of 3.0 [@satmancurcean2016].  
+
+# State of the field
 Nowadays, the interest on statistical computation and data analysis increased the popularity of some programming languages such as R, Python, and Julia. As a result of popularity, hundreds of packages are developed for these targeted languages. On the other side, many mainframe languages are adopted for more specific tasks in server-side or the general-purpose use in several platforms. Interaction between those languages became a major requirement due to the lack of modern statistical tools written in the general-purpose languages such as Java. Renjin [@renjin] is a great effort for interfacing R from within Java using a different perspective: Writing R from scratch in Java! GraalVM [@graalvm] is an other admirable attempt for interfacing Python and R in Java with its polyglot design. Many tools and software packages are developed for interfacing R and Java. Each one stands out with its different features including speed, installation procedures, learning speed, and ease of use. 
 
 
 # Statement of need 
-RCaller is a Java library for interfacing R from within Java [@satman2014rcaller]. R is a popular programming language and a programming environment with hundreds of packages written in C, C++, Fortran, and R itself [@rcoreteam]. These huge collection of computation tools are not directly accessible for the other languages, especially for Java. RCaller supplies a clean API for calling R functions, managing interactions, and transfering objects between languages. There are other options in the literature including Rserve [@urbanek2003fast] and rJava [@urbanek2009talk] which are based on TCP sockets and JNI (Java Native Interface), respectively. RCaller provides a set of easier calling schemes without any dependencies. Previous works showed that the performance of the library is suitable for more cases and studies with moderate datasets can be handled in reasonable times [@satmancurcean2016]. After being introduced in paper [@satman2014rcaller], support for DataFrame objects, R start-up options, automatic Rscript executable locator, and Java Scripting API (JSR 223) are implemented besides minor library changes and performance improvements. RCaller is also used as a core of RCallerService [@rcallerservice], free microservice solution for running R-scripts from any other languages using HTTP.  
[…]
