### 2020-10-31 jbytecode/rcaller#54 @jbytecode
Dear reviewer @xin-huang ,

1. Title changed to "RCaller 3.0: A Java package for interfacing R"

2. Current Summary section renamed to "State of the field".

3. A refactored version of the summary of manual paper is used in the new Summary section.

4. New text added for "Target auidence" as in the README.md.

 Regards
