### 2020-10-31 #2722 @jbytecode
Dear reviewer @xin-huang ,

I think I have completed all the changes you suggested. I also made explanations for each of your suggestions in the issues.

Regards.
