### 2020-10-31 jbytecode/rcaller#54 @xin-huang
Thanks @jbytecode all changes look good
