### 2020-10-31 jbytecode/rcaller#55 @xin-huang
Thanks @jbytecode
I think the changes in the documentation are good.
And please make your benchmark publicly available, then other people could reproduce your results.
