### 2020-10-31 jbytecode/rcaller#55 @jbytecode
Dear reviewer @xin-huang ,

As in the described in the manual paper, I wrote the simulation code from scratch, it is in [PassingArraysAndMatrices.java](scratch). Since it is parameterized, any size of vector and squared matrices can be tested as in the manual and descriptive statistics are reported like

```
- Simulations started.
- Simulations finished. Calculating statistics:
       Min        Max       Mean   Std.Dev.     Median        Mad
 77.000000 445.000000  90.198000  13.782843  89.000000   7.413000
```

  The number of simulations and the number of vector sizes can be changed by the tester. If the number of simulations is reduced then the variance is increased however giving big simulation counts takes more time.

Please tell me if there is anything wrong or missing.

Best regards.
