+++ RCaller/target/apidocs/member-search-index.js
@@ -1,1 +1,0 @@
-memberSearchIndex = [{"p":"com.github.rcaller.util","c":"Globals","l":"about"},{"p":"examples","c":"SampleGui","l":"actionPerformed(ActionEvent)","u":"actionPerformed(java.awt.event.ActionEvent)"},{"p":"com.github.rcaller.util","c":"RCodeUtils","l":"addArray(StringBuilder, String, T[], boolean, boolean)","u":"addArray(java.lang.StringBuilder,java.lang.String,T[],boolean,boolean)"},{"p":"com.github.rcaller.rstuff","c":"RCode","l":"addBoolean(String, boolean)","u":"addBoolean(java.lang.String,boolean)"},{"p":"com.github.rcaller.util","c":"RCodeUtils","l":"addBoolean(StringBuilder, String, boolean, boolean)","u":"addBoolean(java.lang.StringBuilder,java.lang.String,boolean,boolean)"},{"p":"com.github.rcaller.rstuff","c":"RCode","l":"addDataFrame(String, DataFrame)","u":"addDataFrame(java.lang.String,com.github.rcaller.datatypes.DataFrame)"},{"p":"com.github.rcaller.util","c":"RCodeUtils","l":"addDataFrame(StringBuilder, String, DataFrame)","u":"addDataFrame(java.lang.StringBuilder,java.lang.String,com.github.rcaller.datatypes.DataFrame)"},{"p":"com.github.rcaller.rstuff","c":"RCode","l":"addDouble(String, double)","u":"addDouble(java.lang.String,double)"},{"p":"com.github.rcaller.util","c":"RCodeUtils","l":"addDouble(StringBuilder, String, double, boolean)","u":"addDouble(java.lang.StringBuilder,java.lang.String,double,boolean)"},{"p":"com.github.rcaller.rstuff","c":"RCode","l":"addDoubleArray(String, double[])","u":"addDoubleArray(java.lang.String,double[])"},{"p":"com.github.rcaller.util","c":"RCodeUtils","l":"addDoubleArray(StringBuilder, String, double[], boolean)","u":"addDoubleArray(java.lang.StringBuilder,java.lang.String,double[],boolean)"},{"p":"com.github.rcaller.rstuff","c":"RCode","l":"addDoubleMatrix(String, double[][])","u":"addDoubleMatrix(java.lang.String,double[][])"},{"p":"com.github.rcaller.util","c":"RCodeUtils","l":"addDoubleMatrix(StringBuilder, String, double[][], boolean)","u":"addDoubleMatrix(java.lang.StringBuilder,java.lang.String,double[][],boolean)"},{"p":"com.github.rcaller.rstuff","c":"RStreamHandler","l":"addEventHandler(EventHandler)","u":"addEventHandler(com.github.rcaller.EventHandler)"},{"p":"com.github.rcaller.rstuff","c":"RCode","l":"addFloat(String, float)","u":"addFloat(java.lang.String,float)"},{"p":"com.github.rcaller.util","c":"RCodeUtils","l":"addFloat(StringBuilder, String, float, boolean)","u":"addFloat(java.lang.StringBuilder,java.lang.String,float,boolean)"},{"p":"com.github.rcaller.rstuff","c":"RCode","l":"addFloatArray(String, float[])","u":"addFloatArray(java.lang.String,float[])"},{"p":"com.github.rcaller.util","c":"RCodeUtils","l":"addFloatArray(StringBuilder, String, float[], boolean)","u":"addFloatArray(java.lang.StringBuilder,java.lang.String,float[],boolean)"},{"p":"com.github.rcaller.rstuff","c":"RCode","l":"addFunctionCall(Str
[…]
