+++ RCaller/target/apidocs/type-search-index.js
@@ -1,1 +1,0 @@
-typeSearchIndex = [{"l":"All Classes","u":"allclasses-index.html"},{"p":"examples","l":"BasicGraphics"},{"p":"examples","l":"BasicScriptEngine"},{"p":"com.github.rcaller.graphics","l":"BlackTheme"},{"p":"com.github.rcaller.io","l":"CSVFileWriter"},{"p":"com.github.rcaller.datatypes","l":"DataFrame"},{"p":"com.github.rcaller.util","l":"DataFrameUtil"},{"p":"com.github.rcaller.graphics","l":"DefaultTheme"},{"p":"examples","l":"DescriptiveStatistics"},{"p":"examples","l":"EmptyArray"},{"p":"com.github.rcaller","l":"EventHandler"},{"p":"com.github.rcaller.exception","l":"ExecutionException"},{"p":"com.github.rcaller.rstuff","l":"FailurePolicy"},{"p":"com.github.rcaller.util","l":"FileStatusWatcher"},{"p":"examples","l":"Forecasting"},{"p":"com.github.rcaller","l":"FunctionCall"},{"p":"com.github.rcaller","l":"FunctionParameter"},{"p":"com.github.rcaller.util","l":"Globals"},{"p":"com.github.rcaller.graphics","l":"GraphicsTheme"},{"p":"com.github.rcaller.graphics","l":"GraphicsType"},{"p":"examples","l":"GraphicsWithThemes"},{"p":"examples","l":"GridCap"},{"p":"com.github.rcaller","l":"JavaObject"},{"p":"com.github.rcaller.scriptengine","l":"LanguageElement"},{"p":"examples","l":"Main"},{"p":"examples","l":"ManualPaths"},{"p":"com.github.rcaller","l":"MessageSaver"},{"p":"examples","l":"MultipleCalls"},{"p":"com.github.rcaller.scriptengine","l":"NamedArgument"},{"p":"examples","l":"OrdinaryLeastSquares"},{"p":"com.github.rcaller.exception","l":"ParseException"},{"p":"benchmark","l":"PassingArraysAndMatrices"},{"p":"examples","l":"PRandomNumberGenerator"},{"p":"com.github.rcaller.statistics","l":"RandomNumberGenerator"},{"p":"com.github.rcaller.rstuff","l":"RCaller"},{"p":"com.github.rcaller.rstuff","l":"RCallerOptions"},{"p":"com.github.rcaller.scriptengine","l":"RCallerScriptEngine"},{"p":"com.github.rcaller.scriptengine","l":"RCallerScriptEngineFactory"},{"p":"com.github.rcaller.rstuff","l":"RCode"},{"p":"com.github.rcaller.util","l":"RCodeUtils"},{"p":"com.github.rcaller.datatypes","l":"RealMatrix"},{"p":"examples","l":"RegressionWithRandomData"},{"p":"com.github.rcaller.exception","l":"RExecutableNotFoundException"},{"p":"com.github.rcaller.rstuff","l":"ROutputParser"},{"p":"com.github.rcaller.rstuff","l":"RPlotViewer"},{"p":"com.github.rcaller.rstuff","l":"RProcessStartUpOptions"},{"p":"com.github.rcaller.exception","l":"RscriptExecutableNotFoundException"},{"p":"com.github.rcaller.io","l":"RSerializer"},{"p":"com.github.rcaller.rstuff","l":"RService"},{"p":"com.github.rcaller.rstuff","l":"RStreamHandler"},{"p":"examples","l":"SampleGui"},{"p":"examples","l":"SampleRGui"},{"p":"examples","l":"ScriptEngineInvocable"},{"p":"examples","l":"SimplePlot"},{"p":"com.g
[…]
