### 2020-10-31 jbytecode/rcaller#55 @xin-huang
Hi @jbytecode

Because you already have the codes, I think you just highlight it in the README.md, for example, make a link in the README.md to these codes, and let others know if they want to check the performance, they can look at these codes. Thanks.
