### 2020-10-31 jbytecode/rcaller#55 @jbytecode
Dear editor @xin-huang ,

This is mentioned in README.md file [here](https://github.com/jbytecode/rcaller/blob/master/README.md#benchmarks).

Regards.
