### 2020-11-02 #2722 @jbytecode
Dear editor @mikldk ,

1. The term 3.0 was removed from the title.

2 - 3. The compiled pdf seems to be okay.

3. The tagged revision is 3.0.3 and the DOI is [https://zenodo.org/badge/doi/10.5281/zenodo.4189769.svg](https://zenodo.org/badge/doi/10.5281/zenodo.4189769.svg)

The fetched metadata seems okay.
