### 2020-11-03 #2722 @jbytecode
Dear editor @mikldk

- I corrected the authors list and the title in [Zenodo](https://zenodo.org/record/4189769#.X6EYO1MzY_V).

- I made a small correction in the summary, have you got any suggestions for me to correct anything else? It is okay to me.
