### 2020-11-07 #2722 @Kevin-Mattheus-Moerman
@jbytecode I'm helping to process the last steps for this paper. It looks like the paper needs a bit of work to improve language. Can you please consider the points below.

@mikldk could you also help review the paper, since R/Java are not my expertise I want to make sure I did not misunderstood anything.

- [ ] Please add (city and) country information to all affiliations

- [ ] Please review this sentence: "Addition to the previous revisions, RCaller 3.0 implements the scripting API of Java in which the R function calls and data transfers are performed in a standard way as in the way of calling other scripting languages in Java.", it does not read well and the "Addition" part seems incorrect. Consider rephrasing to (if I understood it correctly): "Unlike previous RCaller versions, RCaller 3.0 implements the scripting API of Java, ensuring R function calls and data transfers are performed in a standard way, as is usual for other scripting languages in Java."

- [ ] Check this sentence (use in instead of on): "Nowadays, the interest ~~on~~ in statistical computation and data analysis increased the popularity of some programming languages such as R, Python, and Julia.", also it sounds a bit awkward, would this be better: "Nowadays, increased interest in statistical computation and data analysis has enhanced the popularity of some programming languages such as R, Python, and Julia."

- [ ] Please rephrase this sentence: " On the other side, many main frame languages are adopted for more specific tasks in server-side or the general-purpose use in several platforms."

- [ ] Check this sentence "These huge collection of computation tools are not directly accessible...", it looks like it should probably read "This huge collection of computation tools is not directly accessible..."

- [ ] Rephrase this sentence (perhaps it needs commas): "Previous works showed that the performance of the library is suitable for more cases and studies with moderate data sets can be handled in reasonable times"

- [ ] Consider rephrasing "After being introduced in paper....  are implemented besides ", to something like "Following its first publication.... have been implemented, as well as,... "

- [ ] Suggestion: "matured" -> "mature"

- [ ] Add comma/rephrase: "...and function optimization besides other research tools" -> "...and function optimization, as well as other research tools". Also, if you were to say: "R has many well-tested and mature packages, [including] for...", then you could leave "as well as other research tools" out which sounds less vague.

- [ ] Suggestion "....that R serves... " -> "....that R provides... "

- [ ] "free microservice solution" -> "a free microservice solution".
