### 2020-11-08 20c27a1
paper corrections
