+++ paper/paper.md
@@ -13,9 +13,9 @@ authors:
     orcid: 0000-0002-8223-2168
     affiliation: 2
 affiliations:
- - name: Department of Econometrics, Istanbul University
+ - name: Department of Econometrics, Istanbul University, Istanbul, Turkey
    index: 1
- - name: Department of Mathematical Optimization and Modeling, BIA-Technologies
+ - name: Department of Mathematical Optimization and Modeling, BIA-Technologies, Saint Petersburg, Russian Federation
    index: 2
 date: 24 August 2020
 bibliography: paper.bib
@@ -25,11 +25,14 @@ bibliography: paper.bib
 In this paper, version 3.0 of the open source library RCaller is introduced. RCaller is a software library which simplifies performing data analysis and statistical calculations in Java using R. The details are hidden from users including transferring data between platforms, function calls, and retrieving results. Addition to the previous revisions, RCaller 3.0 implements the scripting API of Java in which the R function calls and data transfers are performed in a standard way as in the way of calling other scripting languages in Java. Besides implementation of new features, RCaller has many performance improvements in the new release of 3.0 [@satmancurcean2016].  
 
 # State of the field
-Nowadays, the interest on statistical computation and data analysis increased the popularity of some programming languages such as R, Python, and Julia. As a result of popularity, hundreds of packages are developed for these targeted languages. On the other side, many mainframe languages are adopted for more specific tasks in server-side or the general-purpose use in several platforms. Interaction between those languages became a major requirement due to the lack of modern statistical tools written in the general-purpose languages such as Java. Renjin [@renjin] is a great effort for interfacing R from within Java using a different perspective: Writing R from scratch in Java! GraalVM [@graalvm] is an other admirable attempt for interfacing Python and R in Java with its polyglot design. Many tools and software packages are developed for interfacing R and Java. Each one stands out with its different features including speed, installation procedures, learning speed, and ease of use. 
+Nowadays, increased interest in statistical computation and data analysis has enhanced the popularity of some programming languages such as R, Python, and Julia. As a result of popularity, hundreds of packages are developed for these targeted languages. On the other side, many main frame languages are adopted for more specific tasks in server-side or the general-purpose use in several platforms. Interaction between those languages became a major requirement due to the lack of modern statistical tools written in the general-purpose languages such as Java. Renjin [@renjin] is a great effort for interfacing R from within Java using a different perspective: Writing R from scratch in Java! GraalVM [@graalvm] is an other admirable attempt for interfacing Python and R in Java with its polyglot design. Many tools and software packages are developed for interfacing R and Java. Each one stands out with its different features including speed, installation procedures, learning speed, and ease of use. 
 
 
 # Statement of need 
[…]
