### 2020-11-08 #2722 @jbytecode
Dear editor @Kevin-Mattheus-Moerman

Thank you very much for your effort on correcting language although you are not obliged to. I made the corrections as suggested above. However, in (2) It is not "unlike" but a "new feature" so other kind of calling schemes are still correct. I changed it to

`In addition to previous RCaller versions, RCaller 3.0 implements the scripting API of Java, ensuring R function calls and data transfers are performed in a standard way, as is usual for other scripting languages in Java.`

Regards.
