### 2020-11-09 #2722 @Kevin-Mattheus-Moerman
@jbytecode I ticked some of the boxes. Some points still remaining. @mikldk can you check some of these sentences and perhaps shed light on how best to rephrase them?
