### 2020-11-09 3be1a96
Update paper.md
This huge collection of computation tools is not directly accessible for the other languages, especially for Java
