### 2020-11-09 #2722 @jbytecode
Dear editor @Kevin-Mattheus-Moerman ,

Sorry to take your precious time for language corrections. Here is my effort to correct unchecked suggestions:
