### 2020-11-09 #2722 @mikldk
@Kevin-Mattheus-Moerman, thanks for catching these. Below you will find my comments to the unchecked boxes:

- [ ] Please rephrase this sentence: " On the other side, many main frame languages are adopted for more specific tasks in server-side or the general-purpose use in several platforms."
  - Change *main frame* to *mainframe* (?) and rephrasing to improve understanding; maybe include the term *domain-specific language*

- [x] Check this sentence "These huge collection of computation tools are not directly accessible...", it looks like it should probably read "This huge collection of computation tools is not directly accessible..."
  - Already fixed from what I can see.

- [x] Add comma/rephrase: "...and function optimization besides other research tools" -> "...and function optimization, as well as other research tools". Also, if you were to say: "R has many well-tested and mature packages, [including] for...", then you could leave "as well as other research tools" out which sounds less vague.
  - Already fixed from what I can see.
