### 2020-11-09 #2722 @jbytecode
Dear editor @mikldk,

The term domain-specific is a little bit a 'constricted concept', like SQL, however I corrected main frame to mainframe
