### 2020-11-10 #2722 @jbytecode
I would like to thank everyone who contributed to the editorial process. It was a great experience to have an article published on JOSS.

All the best.

@Kevin-Mattheus-Moerman @mikldk @xin-huang @nthomasCUBE @Kopilov
